refactor: rename SmallDCliRunnerContext to Conversation and restructure modules
